safari-534.53-branch/LayoutTests/fast/events/event-trace-expected.txt

This tests the target, currentEvent, and phase of a complete trace through event bubbling and capture.

Target = [object HTMLDivElement] (ID: Self); Phase = CAPTURE; CurrentTarget = [object DOMWindow];
Target = [object HTMLDivElement] (ID: Self); Phase = CAPTURE; CurrentTarget = [object HTMLDocument];
Target = [object HTMLDivElement] (ID: Self); Phase = CAPTURE; CurrentTarget = [object HTMLHtmlElement] (ID: HTML);
Target = [object HTMLDivElement] (ID: Self); Phase = CAPTURE; CurrentTarget = [object HTMLBodyElement] (ID: Body);
Target = [object HTMLDivElement] (ID: Self); Phase = CAPTURE; CurrentTarget = [object HTMLDivElement] (ID: Parent);
Target = [object HTMLDivElement] (ID: Self); Phase = AT TARGET; CurrentTarget = [object HTMLDivElement] (ID: Self);
Target = [object HTMLDivElement] (ID: Self); Phase = AT TARGET; CurrentTarget = [object HTMLDivElement] (ID: Self);
Target = [object HTMLDivElement] (ID: Self); Phase = BUBBLE; CurrentTarget = [object HTMLDivElement] (ID: Parent);
Target = [object HTMLDivElement] (ID: Self); Phase = BUBBLE; CurrentTarget = [object HTMLBodyElement] (ID: Body);
Target = [object HTMLDivElement] (ID: Self); Phase = BUBBLE; CurrentTarget = [object HTMLHtmlElement] (ID: HTML);
Target = [object HTMLDivElement] (ID: Self); Phase = BUBBLE; CurrentTarget = [object HTMLDocument];
Target = [object HTMLDivElement] (ID: Self); Phase = BUBBLE; CurrentTarget = [object DOMWindow];

Parent
Self
